+++ yorm/test/test_all_nesting.py
@@ -6,22 +6,51 @@
 from unittest.mock import patch
 
 import pytest
+
 import yorm
 
 from . import strip
 
 
-@yorm.attr(status=yorm.converters.Boolean)
-class NestedDict(yorm.converters.Dictionary):
-    pass
+@yorm.attr(all=yorm.converters.Integer)
+class NestedList2(yorm.converters.List):
+
+    def __repr__(self):
+        return "<nested-list-2 {}>".format((id(self)))
 
 
-@yorm.attr(all=NestedDict)
+@yorm.attr(number=yorm.converters.Integer)
+class NestedDictionary2(yorm.converters.AttributeDictionary):
+
+    def __init__(self):
+        super().__init__()
+        self.number = 0
+
+    def __repr__(self):
+        return "<nested-dictionary-2 {}>".format((id(self)))
+
+
+@yorm.attr(number=yorm.converters.Integer)
+@yorm.attr(numbers=NestedList2)
+class NestedDictionary(yorm.converters.AttributeDictionary):
+
+    def __init__(self):
+        super().__init__()
+        self.number = 0
+        self.numbers = []
+
+    def __repr__(self):
+        return "<nested-dictionary {}>".format((id(self)))
+
+
+@yorm.attr(all=NestedDictionary2)
 class NestedList(yorm.converters.List):
-    pass
+
+    def __repr__(self):
+        return "<nested-list {}>".format((id(self)))
 
 
-@yorm.attr(nested_list=NestedList, nested_dict=NestedDict)
+@yorm.attr(nested_list=NestedList, nested_dict=NestedDictionary)
 @yorm.sync("sample.yml")
 class Top:
 
@@ -29,44 +58,82 @@ class Top:
         self.nested_list = []
         self.nested_dict = {}
 
+    def __repr__(self):
+        return "<top {}>".format((id(self)))
+
 
 @patch('yorm.settings.fake', True)
 class TestNesting:
 
     def test_list_append_triggers_update(self):
         top = Top()
-        top.nested_list.append(None)
+        top.nested_list.append({'number': 1})
         assert strip("""
         nested_dict:
-          status: false
+          number: 0
+          numbers: []
         nested_list:
-        - status: false
+        - number: 1
+        """) == top.yorm_mapper.text
+
+    def test_list_append_triggers_update2(self):
+        top = Top()
+        top.nested_list.append({'number': 1})
+        assert strip("""
+        nested_dict:
+          number: 0
+          numbers: []
+        nested_list:
+        - number: 1
         """) == top.yorm_mapper.text
 
     def test_dict_set_triggers_update(self):
         top = Top()
-        top.nested_dict['status'] = 1
+        top.nested_dict.number = 2
         assert strip("""
         nested_dict:
-          status: true
+          number: 2
+          numbers: []
         nested_list: []
         """) == top.yorm_mapper.text
 
-    def test_dict_in_list_triggers_update(self):
+    def test_list_item_value_change_triggers_update(self):
         top = Top()
-        top.nested_list = {'status': True}
+        top.nested_list = [{'number': 3}]
         assert strip("""
         nested_dict:
-          status: false
+          number: 0
+          numbers: []
         nested_list:
-        - status: true
+        - number: 3
         """) == top.yorm_mapper.text
-        top.nested_list[0]['status'] = False
+        top.nested_list[0].number = 4
         assert strip("""
         nested_dict:
-          status: false
+          number: 0
+          numbers: []
         nested_list:
-        - status: false
+        - number: 4
+        """) == top.yorm_mapper.text
+
+    def test_dict_item_value_change_triggers_update(self):
+        top = Top()
+        top.nested_dict = {'numbers': [5]}
+        assert strip("""
+        nested_dict:
+          number: 0
+          numbers:
+          - 5
+        nested_list: []
+        """) == top.yorm_mapper.text
+        top.nested_dict.numbers.append(6)
+        assert strip("""
+        nested_dict:
+          number: 0
+          numbers:
+          - 5
+          - 6
+        nested_list: []
         """) == top.yorm_mapper.text