+++ yorm/utilities.py
@@ -45,14 +45,12 @@ def sync_object(instance, path, attrs=None, auto=True):
 
         """Original class with `Mappable` as the base."""
 
-    mapper = Mapper(instance, path, attrs)
+    mapper = Mapper(instance, path, attrs, auto=auto)
+
     if not mapper.exists:
         mapper.create()
-        if auto:
-            mapper.store()
-    else:
-        mapper.fetch()
-    mapper.auto = auto
+        mapper.store()
+    mapper.fetch(force=True)
 
     setattr(instance, MAPPER, mapper)
     instance.__class__ = Mapped
@@ -92,17 +90,14 @@ def sync_instances(path_format, format_spec=None, attrs=None, auto=True):
                 path = path_format.format(**format_values)
                 attrs.update(common.ATTRS[self.__class__])
                 attrs.update(common.ATTRS[cls])
+                mapper = Mapper(self, path, attrs, auto=auto)
 
-                mapper = Mapper(self, path, attrs)
                 if not mapper.exists:
                     mapper.create()
-                    if auto:
-                        mapper.store()
-                else:
-                    mapper.fetch()
-                mapper.auto = auto
+                    mapper.store()
+                mapper.fetch(force=True)
 
-                self.yorm_mapper = mapper
+                setattr(self, MAPPER, mapper)
 
         return Mapped
 
@@ -151,9 +146,11 @@ def update_object(instance, force=True):
     :param force: update the object even if the file appears unchanged
 
     """
+    log.info("manually synchronizing %r from its file...", instance)
     _check_base(instance, mappable=True)
 
-    instance.yorm_mapper.fetch(force=force)
+    mapper = getattr(instance, MAPPER)
+    mapper.fetch(force=force)
 
 
 def update_file(instance):
@@ -162,9 +159,11 @@ def update_file(instance):
     :param instance: object with patched YAML mapping behavior
 
     """
+    log.info("manually synchronizing %r to its file...", instance)
     _check_base(instance, mappable=True)
 
-    instance.yorm_mapper.store()
+    mapper = getattr(instance, MAPPER)
+    mapper.store()
 
 
 def _check_base(obj, mappable=True):