+++ CHANGES.md
@@ -7,6 +7,7 @@ Changelog
 - Moved all converters into the `yorm.converters` package.
 - Rename `Converter` to `Convertible`.
 - Remove the context manager in mapped objects.
+- Fixed automatic mapping of nested attributes.
 
 0.3.2 (2015-04-07)
 ------------------